+++ graphics/r_graphics.cpp
@@ -35,7 +35,7 @@ bool Graphics::init() {
     
     //: Create renderer api
     API::createShaderList();
-    API::renderer_description_path = "res/render/renderer_description";
+    API::renderer_description_path = "render/renderer_description";
     api = API::createAPI(win);
 
     //: Set projection