+++ CHANGELOG.md
@@ -6,10 +6,15 @@
 
 #### 0.3.11 - instanced rendering - _14 03 2022_
 
+**added**
+- instanced rendering!
+
 **changed**
 - multiple vertex types in attribute description (to allow instance rendering vertices)
 - complete rework of the rendering queue and rendering data
-- instanced rendering working!
+
+**fixed**
+- rendering description not loading on web due to path issue
 
 ---